- Use only one environment config while will then be populated depending   on passed configuration file. - Remove private server URLs in the configuration file.
